+++ library/CMakeLists.txt
@@ -27,7 +27,6 @@ set(src_crypto
     cipher.c
     cipher_wrap.c
     cmac.c
-    common.c
     ctr_drbg.c
     des.c
     dhm.c
